+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
+### 0.8.22
+  - Fixes ember-metal related issues on Ember Canary.
+
+### 0.8.21
+  - Fixes ES5 getter syntax on Ember Canary (#248)
+
 ### 0.8.18
   - Adds `.evented()` task modifier and task lifecycle events.